refactor(e2e): folder change for annotations.ts
