+++ packages/suite-desktop-core/e2e/support/reporters/annotations.ts
@@ -13,7 +13,7 @@ import {
     annotationsAddedToTest,
     annotationsForBodyDescription,
     annotationsForProjectFields,
-} from './enums/testAnnotations';
+} from '../enums/testAnnotations';
 
 type TestMetadataInput = {
     testCase?: string;
