+++ packages/suite-desktop-core/e2e/tests/backup/t3t1-create-additional-share.test.ts
@@ -1,8 +1,8 @@
 import { MNEMONICS } from '@trezor/trezor-user-env-link';
 
-import { createTestAnnotation } from '../../support/annotations';
 import { TestCategory, TestPriority, TestStream } from '../../support/enums/testAnnotations';
 import { test } from '../../support/fixtures';
+import { createTestAnnotation } from '../../support/reporters/annotations';
 
 test.describe('Create additional share', { tag: ['@group=device-management'] }, () => {
     test.use({
