+++ packages/suite-desktop-core/e2e/tests/manual/cardano/cardano-staking.test.ts
@@ -1,6 +1,6 @@
-import { createTestAnnotation } from '../../../support/annotations';
 import { TestCategory, TestPriority, TestStream } from '../../../support/enums/testAnnotations';
 import { test } from '../../../support/fixtures';
+import { createTestAnnotation } from '../../../support/reporters/annotations';
 
 test.describe.skip('Cardano staking', { tag: ['@group=manual'] }, () => {
     test(
