+++ packages/suite-desktop-core/e2e/tests/manual/firmware/firmware-update.test.ts
@@ -1,6 +1,6 @@
-import { createTestAnnotation } from '../../../support/annotations';
 import { TestCategory, TestPriority, TestStream } from '../../../support/enums/testAnnotations';
 import { test } from '../../../support/fixtures';
+import { createTestAnnotation } from '../../../support/reporters/annotations';
 
 test.describe.skip('Firmware update', { tag: ['@group=manual'] }, () => {
     test(
