+++ packages/suite-desktop-core/e2e/tests/manual/settings/check-update.test.ts
@@ -1,6 +1,6 @@
-import { createTestAnnotation } from '../../../support/annotations';
 import { TestCategory, TestPriority } from '../../../support/enums/testAnnotations';
 import { test } from '../../../support/fixtures';
+import { createTestAnnotation } from '../../../support/reporters/annotations';
 
 test.describe.skip('Check for update', { tag: ['@group=manual'] }, () => {
     test(
