+++ packages/suite-desktop-core/e2e/tests/manual/settings/wallet-loading.test.ts
@@ -1,6 +1,6 @@
-import { createTestAnnotation } from '../../../support/annotations';
 import { TestCategory, TestPriority } from '../../../support/enums/testAnnotations';
 import { test } from '../../../support/fixtures';
+import { createTestAnnotation } from '../../../support/reporters/annotations';
 
 test.describe.skip('Wallet loading', { tag: ['@group=manual'] }, () => {
     test(
